requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
python-dotenv==1.0.0
requests==2.30.0
aiohttp==3.8.4
httpx==0.24.0
asyncio==3.4.3
python-multipart==0.0.6
strawberry-graphql==0.193.1

# Database and cache
pymongo==4.3.3
redis==4.5.5
motor==3.1.2
beanie==1.19.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
aiomysql==0.1.1

# Authentication and security
pyjwt==2.7.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
cryptography==40.0.2

# Machine learning and predictive analytics
xgboost==1.7.5
lightgbm==3.3.5
catboost==1.2
statsmodels==0.14.0
pmdarima==2.0.3
prophet==1.1.4
bayesian-optimization==1.4.3
scipy==1.10.1
joblib==1.2.0

# NLP components for factor parsing
nltk==3.8.1
spacy==3.5.2
transformers==4.29.2
sentence-transformers==2.2.2
gensim==4.3.1
textdistance==4.5.0
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# Visualization (for analytics)
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1

# Monitoring and logging
prometheus-client==0.16.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-instrumentation-fastapi==0.39b0
loguru==0.7.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-mock==3.10.0
pytest-cov==4.1.0
faker==18.6.0

# Documentation
sphinx==6.2.1
sphinx-rtd-theme==1.2.0
mkdocs==1.4.3
mkdocs-material==9.1.8

# Utils
tqdm==4.65.0
tenacity==8.2.2
backoff==2.2.1
orjson==3.8.12
pyyaml==6.0
fastapi-limiter==0.1.5
colorlog==6.7.0

# Download required spaCy models
# python -m spacy download en_core_web_sm
# python -m nltk.downloader punkt stopwords